clar: pass a string for a `%s` format placeholder
When `clar_parse_args()` encounters an unhandled command-line option, it
wants to print out an error message that says so, and show the option
that it encountered. However, the format claims that it wants to show a
string, whereas using `argument[1]` provided only the first character.

This was side-stepped in upstream `clar` in 7eaea8b2da (clar: argument
parsing improvements, 2025-01-16) by not even showing the option
anymore.

Let's just fix this in Git until the next time Git synchronizes with an
upstream `clar` version again.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>