release-winget.yml: fix tag name on releases
When running this workflow from the `release` event the parsing of the
tag name from the event object must happen _after_ the object has been
loaded, not before! This error eventually results in a bogus value for
the `$version` argument that gets passed to `wingetcreate`.

Ensure we only access the `$github` variable _after_ we've defined it.

Signed-off-by: Matthew John Cheetham <mjcheetham@outlook.com>