fixup! release: create initial Windows installer build workflow
As pointed out in https://github.com/microsoft/git/issues/739, the VS
registry settings are added only if the autoupdater feature is enabled.
This was unintentional; Just like the corresponding part in the
uninstaller, these two things should be apart.

So let's decouple the VS registry settings from the autoupdater.

Let's do that by ensuring that we put the function calls _outside_ of
the autoupdater blocks.

For good measure, we add checks to verify that both calls _were_ added,
i.e. that the expected patterns were found.

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>