fixup! .github/actions/akv-secret: add action to get secrets
Use a buffer rather than a string when handling the output from the `az`
command. Handling of binary data (that was base64 encoded) requires that
we use a buffer and not a string, or else we will end up writing invalid
data to files/output variables.

Introduce several new helper functions for working with buffers,
including trimming the EOL bytes (CR, LF), and fix up some of the
output functions to correctly validate the value passed - it must be
something printable (UTF-8-ish).

Finally ensure that we correctly mask multi-line secret values by
emitting a `::add-mask` command for each line.

Co-authored-by: Johannes Schindelin <johannes.schindelin@gmx.de>
Signed-off-by: Matthew John Cheetham <mjcheetham@outlook.com>