ci(jgit): use a more reliable link to download JGit
Git's dependency story is improvable: When, say, JGit cannot be
downloaded, it currently fails the entire job without even running all
the tests that do not require JGit (which is the vast, vast majority of
them). It would be better to handle that gracefully by warning about the
JGit download failure and then skipping the respective tests.

This JGit download problem is far from theoretical: There is an incident
on eclipse.org that has now lasted for an extended amount of
time: https://www.eclipsestatus.io/incident/549796.

While upstream Git finally chose to improve their CI story in
b0026daf1e60 (ci: skip unavailable external software, 2025-04-24), we
should heed the suggestion by the JGit maintainer, Matthias Sohn, in
https://discord.com/channels/1042895022950994071/1364872237710184520/1364885895865696479,
to switch to the far more reliable Maven Central link (according to
https://status.maven.org/uptime, there has not been any major downtime
recently that wasn't resolved within 5 minutes, a far cry from the 30h
of the current eclipse.org incident as of time of writing).

Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>