hook: restrict post-command hook even more
The postCommand.strategy=post-index-change config was designed to
replace some internal hook usage. This internal hook was more
restrictive than the initial feature was designed. Specifically, it
ignored index updates that did not update the worktree. With the
existing implementation, this causes more work done in the post-command
hook than intended.

Update the code to write a sentinel file only when the worktree is
changed.

Signed-off-by: Derrick Stolee <stolee@gmail.com>