hooks: change postCommand.strategy config value
The previously-named 'post-index-change' value for
'postCommand.strategy' did the following two things:

 1. Avoid running the post-command hook unless the post-index-change
    hook _would_ run with a signal that the worktree changed.

 2. Avoid running any installed post-index-change hooks.

The additional restriction of the worktree change in (1) makes this name
less appropriate. Also, item (2) seems to be a side-effect that we
should avoid. We would like to allow both behaviors to exist.

Update the value and behavior, including the tests.

Signed-off-by: Derrick Stolee <stolee@gmail.com>