scalar: do use HTTP/1.1 to check whether GVFS is supported
The way the `gvfs-helper` command operates is apparently incompatible
with HTTP/2, that's why https://github.com/microsoft/git/pull/754 tried
to enforce HTTP/1.1 in Scalar clones accessing Azure Repos.

However, this fix was incomplete for the `scalar clone` part because it
made the HTTP/1.1 enforcement contingent on the test whether the remote
repository supports the GVFS protocol or not, and that test did _not_
enforce HTTP/1.1.

Let's fix that Catch-22 by enforcing HTTP/1.1 usage during that GVFS
Protocol capability test.

Reported-by: Derrick Stolee <stolee@gmail.com>
Signed-off-by: Johannes Schindelin <johannes.schindelin@gmx.de>